move loadPagedIDs() to util.go
